Allow iodata for log message input and tolerate poorly formed messages.
If bad messages come in at a high enough rate, it can cause the supervisor to give up,
which brings down the whole app. Given the conditional nature of logging it is difficult
to test all log statements in an application, so I'd personally rather the app stay
running than crash the app for the logging case. We may want to do something in the catch/2
to indicate that something went wrong with the msg format, but this commit doesn't
address that. Maybe log a subset of the metadata indicating where the problem may be,
because we certainly can't log the msg.